Don't enable HidDevService if already enabled
In #3, com.android.bluetooth is constantly restarting and the hook is
repeatedly being called. I suspect that the DONT_KILL_APP flag for
setComponentEnabledSetting isn't actually working, so this commit
changed behaviour to only force enable if it isn't already.

Note that getComponentEnabledSetting returns
COMPONENT_ENABLED_STATE_DEFAULT if it hasn't been force enabled or
disabled - we still force enable even if it's enabled by default.